Add UtilisateurController implementation and supprimerCategorie method in CategorieController
